+++ tests/core/test_circuit.py
@@ -401,7 +401,7 @@ def test_tensor(
         (np.array([1, 1j]) / np.sqrt(2)),
     ],
 )
-def test_initializer(state: npt.NDArray[np.complex64]):
+def test_initializer(state: npt.NDArray[np.complex128]):
     qc = QCircuit.initializer(state)
     res = run(qc, IBMDevice.AER_SIMULATOR_STATEVECTOR)
     if TYPE_CHECKING:
@@ -441,7 +441,7 @@ def test_count(circuit: QCircuit, filter: tuple[type[Gate]], count: int):
                 [
                     BasisMeasure([0, 1], shots=1000),
                     ExpectationMeasure(
-                        Observable(np.identity(2, dtype=np.complex64)), [1], shots=1000
+                        Observable(np.identity(2, dtype=np.complex128)), [1], shots=1000
                     ),
                 ]
             ),
